Reorganized the prototype functions a bit
The operator overloading prototypes were turned into friends of the classes they are overloaded for. This makes the code in the header a little bit more readable since the operator overloadings are not listed one after another along with the other function prototypes that don't belong to one or more classes.